@@ -1,23 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-
-// import the browser router
-import {BrowserRouter} from 'react-router-dom'
-import {FavoritesContextProvider} from './store/favorites-context'
-
-
-
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(
-  <FavoritesContextProvider>
-      <BrowserRouter>
-        <App />
-      </BrowserRouter>
-  </FavoritesContextProvider>,
-  document.getElementById('root')
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
